requirements.txt
django-rest-framework
django
django-cors-headers